Backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-decouple
Pillow
django-filter
